# Conflicts:
#	Application-Code/app/build/intermediates/incremental/mergeDebugResources/compile-file-map.properties
#	Application-Code/app/build/intermediates/incremental/packageDebug/tmp/debug/dex-renamer-state.txt
#	Application-Code/app/build/intermediates/javac/debug/classes/com/example/medication_reminder_android_app/SQLiteDB/DataAccessObject_Impl$10.class
#	Application-Code/app/build/intermediates/javac/debug/classes/com/example/medication_reminder_android_app/SQLiteDB/DataAccessObject_Impl$11.class
#	Application-Code/app/build/intermediates/javac/debug/classes/com/example/medication_reminder_android_app/SQLiteDB/DataAccessObject_Impl$12.class
#	Application-Code/app/build/intermediates/javac/debug/classes/com/example/medication_reminder_android_app/SQLiteDB/DataAccessObject_Impl$9.class
#	Application-Code/app/build/intermediates/javac/debug/classes/com/example/medication_reminder_android_app/SQLiteDB/DatabaseRepository$AsyncDeleteAllApptReminders.class
#	Application-Code/app/build/intermediates/javac/debug/classes/com/example/medication_reminder_android_app/SQLiteDB/DatabaseRepository$AsyncDeleteAllMedReminders.class
#	Application-Code/app/build/intermediates/javac/debug/classes/com/example/medication_reminder_android_app/SQLiteDB/DatabaseRepository$AsyncGetMedById.class
#	Application-Code/app/build/intermediates/javac/debug/classes/com/example/medication_reminder_android_app/SQLiteDB/DatabaseRepository$AsyncGetReminderById.class
#	Application-Code/app/build/intermediates/javac/debug/classes/com/example/medication_reminder_android_app/SQLiteDB/DatabaseRepository$AsyncReminderUpdateDateTime.class
#	Application-Code/app/build/intermediates/javac/debug/classes/com/example/medication_reminder_android_app/SQLiteDB/DatabaseRepository$AsyncUpdateAcknowledgement.class
#	Application-Code/app/build/intermediates/project_dex_archive/debug/out/com/example/medication_reminder_android_app/SQLiteDB/DataAccessObject_Impl$10.dex
#	Application-Code/app/build/intermediates/project_dex_archive/debug/out/com/example/medication_reminder_android_app/SQLiteDB/DataAccessObject_Impl$11.dex
#	Application-Code/app/build/intermediates/project_dex_archive/debug/out/com/example/medication_reminder_android_app/SQLiteDB/DataAccessObject_Impl$12.dex
#	Application-Code/app/build/intermediates/project_dex_archive/debug/out/com/example/medication_reminder_android_app/SQLiteDB/DataAccessObject_Impl$9.dex
#	Application-Code/app/build/intermediates/project_dex_archive/debug/out/com/example/medication_reminder_android_app/SQLiteDB/DatabaseRepository$AsyncDeleteAllApptReminders.dex
#	Application-Code/app/build/intermediates/project_dex_archive/debug/out/com/example/medication_reminder_android_app/SQLiteDB/DatabaseRepository$AsyncDeleteAllMedReminders.dex
#	Application-Code/app/build/intermediates/project_dex_archive/debug/out/com/example/medication_reminder_android_app/SQLiteDB/DatabaseRepository$AsyncGetMedById.dex
#	Application-Code/app/build/intermediates/project_dex_archive/debug/out/com/example/medication_reminder_android_app/SQLiteDB/DatabaseRepository$AsyncGetReminderById.dex
#	Application-Code/app/build/intermediates/project_dex_archive/debug/out/com/example/medication_reminder_android_app/SQLiteDB/DatabaseRepository$AsyncReminderUpdateDateTime.dex
#	Application-Code/app/build/intermediates/project_dex_archive/debug/out/com/example/medication_reminder_android_app/SQLiteDB/DatabaseRepository$AsyncUpdateAcknowledgement.dex

+++ Application-Code/app/src/main/java/com/example/medication_reminder_android_app/NotificationRelay/Notifications.java
@@ -17,7 +17,7 @@ public class Notifications {
 
     }
 
-         /*
+    /**
     @author: Karley Waguespack
     Last Modified: 03/22/2021
 
@@ -38,7 +38,7 @@ public class Notifications {
      */
 
     //should take either a reminder or a reminder ID
-    protected String[] getData(Integer reminderID) {
+    protected String[] getData(long reminderID) {
 
         //string info array to be returned
         String[] infoArray = new String[2];
@@ -62,7 +62,7 @@ public class Notifications {
 
         }
 
-        infoArray[2] = Integer.toString(reminderID);
+        infoArray[2] = Long.toString(reminderID);
         return infoArray;
     }
